chore(pane metric) Extract PaneMetricInfoContent component
