+++ src/lib/ui/app/Chart/PaneLegend/index.ts
@@ -4,6 +4,7 @@ export { default as PaneMetricValue } from './Metric/Value.svelte'
 export { default as PaneMetricGranularityStatus } from './Metric/GranularityStatus.svelte'
 export { default as PaneMetricVersionStatus } from './Metric/VersionStatus.svelte'
 export { default as PaneMetricSettingsSchemaStatus } from './Metric/SettingsSchemaStatus.svelte'
+export { default as PaneMetricInfoContent } from './Metric/InfoContent.svelte'
 export { default as AIExplanationStatus } from './Metric/AIExplanationStatus.svelte'
 
 export { usePaneLegendCompactCtx } from './ctx.svelte.js'
